prepare Breadcrumbs to be split into multible components
